# bpo-29585: optimize site.py startup time