--- a/pyconfig.h.in
+++ b/pyconfig.h.in
@@ -1247,6 +1247,9 @@
 /* Define as the preferred size in bits of long digits */
 #undef PYLONG_BITS_IN_DIGIT
 
+/* framework name */
+#undef PYTHONFRAMEWORK
+
 /* Define if you want to coerce the C locale to a UTF-8 based locale */
 #undef PY_COERCE_C_LOCALE