@methane 2017-02-18
I've addes `sys._framework` and remove sysconf dependency completely.

As suggested [devguide](http://cpython-devguide.readthedocs.io/setup.html#regenerate-configure),
I didn't commit autogenerated files for now.
I'll commit them when got consensus.