@methane 2017-06-28
@ned-deily I created bpo issue and added comment about duplicated code.