@methane 2017-06-28
@haypo sysconfig (and `_sysconfigdata_...`) module is relatively large and
most applications (except packaging tools) doesn't use it at all.
Now this pull request is focusing it.

Slow code path is skipped by #167 although `abspath` is still slow.