@1st1 2017-06-28
Why couldn't we make 'sysconfig' to import a couple of private functions from 'site', why copying?