@haypo 2017-06-28
> Thank you for review.

Thanks for this cool and cheap speedup!