@haypo 2017-06-28
Oh, Windows build is broken :-(

```
     2>..\Python\sysmodule.c(1968): error C2065: 'PYTHONFRAMEWORK': undeclared identifier [C:\buildbot.python.org\3.x.kloth-win64\build\PCbuild\pythoncore.vcxproj]
```

Need to fix PC/pyconfig.h?