@haypo 2017-06-28
Oh, AppVeyor didn't catch the bug simply because it wasn't run on this PR!