@haypo 2017-06-28
@methane proposed PR #2476 to fix PC/pyconfig.h, I proposed PR #2477.